Extract TaskAlreadyClaimedException into common module
This change removes the usage of org.flowable.rest.exception from the flowable-rest module
