+++ modules/flowable-cmmn-rest/src/test/java/org/flowable/rest/cmmn/conf/common/RestExceptionHandlerAdvice.java
@@ -12,12 +12,12 @@
  */
 package org.flowable.rest.cmmn.conf.common;
 
-import org.flowable.rest.cmmn.exception.CaseExceptionHandlerAdvice;
+import org.flowable.rest.exception.BaseExceptionHandlerAdvice;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 
 /**
  * @author Yvo Swillens
  */
 @ControllerAdvice
-public class RestExceptionHandlerAdvice extends CaseExceptionHandlerAdvice {
+public class RestExceptionHandlerAdvice extends BaseExceptionHandlerAdvice {
 }
