+++ modules/flowable-engine-common-api/src/main/java/org/flowable/engine/common/api/FlowableTaskAlreadyClaimedException.java
@@ -10,9 +10,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.flowable.cmmn.api.exception;
-
-import org.flowable.engine.common.api.FlowableException;
+package org.flowable.engine.common.api;
 
 /**
  * This exception is thrown when you try to claim a task that is already claimed by someone else.
