+++ modules/flowable-engine/src/main/java/org/flowable/engine/TaskService.java
@@ -152,8 +152,8 @@ public interface TaskService {
      *            user that claims the task. When userId is null the task is unclaimed, assigned to no one.
      * @throws FlowableObjectNotFoundException
      *             when the task doesn't exist.
-     * @throws FlowableTaskAlreadyClaimedException
-     *             when the task is already claimed by another user.
+     * @throws org.flowable.engine.common.api.FlowableTaskAlreadyClaimedException
+     *             when the task is already claimed by another user
      */
     void claim(String taskId, String userId);
